# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PTO / Vacation Calendar API
  version: 1.1.0
  description: |
    API for managing PTO requests in a shared team calendar.

    ## Operational notes
    - Every response carries an `X-Request-Id` header (UUID v4) that mirrors
      an inbound `X-Request-Id` when the client supplied one. Use this to
      correlate with server logs.
    - Two unauthenticated probes are exposed for orchestrators: `GET /health`
      (liveness, no DB) and `GET /ready` (readiness, DB probe).
    - All routes are subject to a global IP rate limiter
      (`RATE_LIMIT_MAX` per `RATE_LIMIT_WINDOW_MS`). `POST /auth/login` is
      additionally subject to a stricter login limiter
      (`AUTH_RATE_LIMIT_MAX` per window) that counts failed attempts only.
      Both return `429 Too Many Requests` with `Retry-After` and the body
      `{ "error": { "code": "RATE_LIMITED", "message": "..." } }`.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local
security:
  - cookieAuth: []
paths:
  /health:
    get:
      security: []
      summary: Liveness probe
      tags: [Health]
      description: |
        Always returns 200 with `{ "status": "ok" }` while the process is
        responsive. Does not check the database. Use this to decide whether
        to restart the process.
      responses:
        '200':
          description: Service is alive
          headers:
            X-Request-Id:
              description: Correlation ID echoed from the request, or a generated UUID v4.
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
  /ready:
    get:
      security: []
      summary: Readiness probe
      tags: [Health]
      description: |
        Runs `SELECT 1` against the database with a `READY_TIMEOUT_MS` race
        timer. Returns 200 when the DB responds in time, 503 otherwise.
        Use this to decide whether to route traffic.
      responses:
        '200':
          description: Service is ready to serve traffic
          headers:
            X-Request-Id:
              description: Correlation ID echoed from the request, or a generated UUID v4.
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                type: object
                required: [status, db, uptime]
                properties:
                  status:
                    type: string
                    enum: [ready]
                  db:
                    type: string
                    enum: [ok]
                  uptime:
                    type: number
                    description: Process uptime in seconds.
        '503':
          description: Service is not ready (DB unreachable or probe timed out)
          headers:
            X-Request-Id:
              description: Correlation ID echoed from the request, or a generated UUID v4.
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notReady:
                  value:
                    error:
                      code: NOT_READY
                      message: 'Database is not reachable'
                      details:
                        db: unreachable
                        reason: timeout
                        timeoutMs: 5000
  /auth/login:
    post:
      security: []
      summary: Log in a user
      tags: [Auth]
      description: |
        Subject to the strict login limiter (`AUTH_RATE_LIMIT_MAX` failed
        attempts per `RATE_LIMIT_WINDOW_MS` per IP, with
        `skipSuccessfulRequests: true`). Also subject to the global limiter.
        On rate-limit hit, returns `429` with `Retry-After` and the body
        `{ "error": { "code": "RATE_LIMITED", "message": "..." } }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in successfully
          headers:
            X-Request-Id:
              description: Correlation ID echoed from the request, or a generated UUID v4.
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests (login or global limiter tripped)
          headers:
            Retry-After:
              description: Seconds until the limiter window resets.
              schema: { type: integer, minimum: 0 }
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rateLimited:
                  value:
                    error:
                      code: RATE_LIMITED
                      message: 'Too many login attempts. Please try again later.'
  /auth/logout:
    post:
      summary: Log out current user
      tags: [Auth]
      responses:
        '204':
          description: Logged out
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
  /auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      responses:
        '200':
          description: Current user
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/setup-account:
    post:
      summary: Redeem a one-time setup token and set the initial password
      tags: [Auth]
      description: |
        Public endpoint. The lead creates a user via `POST /users` and shares
        the returned one-time `setupToken` link with the new member; the member
        opens `/setup-account?token=...` and submits a password. The token is
        sha256-hashed in the database, expires after `SETUP_TOKEN_TTL_MS`
        (default 24h), and is single-use. Rate-limited by the same login
        limiter as `POST /auth/login`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupAccountRequest'
      responses:
        '200':
          description: Password set, session established, user returned.
          headers:
            Set-Cookie:
              description: session cookie issued via `cookie-session`.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error (token or password missing/invalid).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Token is missing/expired/already-spent. Identical response to avoid enumeration.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          description: Setup-account rate limit tripped.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users:
    get:
      summary: List users (team lead only)
      tags: [Users]
      description: Returns the public user list. Requires team lead.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Authenticated but not a team lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      summary: Create a user (team lead only)
      tags: [Users]
      description: |
        Team-lead only. Creates a member with `passwordHash: null` and a fresh
        one-time setup token; returns the user + plaintext `setupToken` +
        `expiresAt`. The setup link is shown ONCE in the admin UI for the lead
        to copy.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Authenticated but not a team lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Email already exists (lowercased uniqueness).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{id}/reset-password:
    post:
      summary: Reset a user's password (team lead only)
      tags: [Users]
      description: |
        Team-lead only. Clears the target user's `passwordHash`, issues a fresh
        one-time setup token, and returns the new plaintext token. Blocks
        self-reset and reset of the only team_lead in the system.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Reset token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResponse'
        '400':
          description: Attempted to reset the only team_lead (refused) or reset self.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Authenticated but not a team lead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /pto:
    get:
      summary: List PTO entries overlapping a visible calendar date range
      tags: [PTO]
      description: Subject to the global rate limiter; may return `429`.
      parameters:
        - in: query
          name: start
          required: true
          schema:
            type: string
            format: date
            example: 2026-04-27
          description: Visible calendar grid start date in YYYY-MM-DD format
        - in: query
          name: end
          required: true
          schema:
            type: string
            format: date
            example: 2026-05-31
          description: Visible calendar grid end date in YYYY-MM-DD format
      responses:
        '200':
          description: PTO entries overlapping the requested visible calendar date range
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PTOWithUser'
        '400':
          description: Invalid date range parameter
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      summary: Create PTO entry for current user
      tags: [PTO]
      description: Subject to the global rate limiter; may return `429`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePTORequest'
      responses:
        '201':
          description: PTO created
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTO'
        '400':
          description: Validation error
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: PTO overlaps an existing PTO entry for the same user
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /pto/{id}:
    get:
      summary: Get PTO by id
      tags: [PTO]
      description: Subject to the global rate limiter; may return `429`.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PTO details
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTOWithUser'
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: PTO not found
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
    put:
      summary: Update PTO by id
      tags: [PTO]
      description: Subject to the global rate limiter; may return `429`.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePTORequest'
      responses:
        '200':
          description: PTO updated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PTO'
        '400':
          description: Validation error
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not allowed to modify PTO
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: PTO not found
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      summary: Delete PTO by id
      tags: [PTO]
      description: Subject to the global rate limiter; may return `429`.
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: PTO deleted
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
        '401':
          description: Not authenticated
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not allowed to delete PTO
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: PTO not found
          headers:
            X-Request-Id:
              schema: { type: string, format: uuid }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  responses:
    RateLimited:
      description: |
        Too many requests. Returned by the global limiter on every route and
        by the stricter login limiter on `POST /auth/login`.
      headers:
        Retry-After:
          description: Seconds until the limiter window resets.
          schema: { type: integer, minimum: 0 }
        X-Request-Id:
          description: Correlation ID echoed from the request, or a generated UUID v4.
          schema: { type: string, format: uuid }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rateLimited:
              value:
                error:
                  code: RATE_LIMITED
                  message: 'Too many requests. Try again later.'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: Optional client-supplied correlation ID; echoed in the response header and used in logs.
      schema:
        type: string
        format: uuid
  schemas:
    Role:
      type: string
      enum: [member, team_lead, admin]
    DayPart:
      type: string
      enum: [morning, evening, all_day]
    User:
      type: object
      required: [id, name, email, role, colorCode]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        colorCode:
          type: string
          example: '#3B82F6'
    PTO:
      type: object
      required: [id, userId, startDate, endDate, dayPart, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dayPart:
          $ref: '#/components/schemas/DayPart'
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PTOWithUser:
      type: object
      description: >
        Returned by GET /pto (list) and GET /pto/:id (detail). The list response
        always has `note: null`; the detail response includes `note` for the PTO
        owner or a team lead, and `null` otherwise.
      required: [id, user, startDate, endDate, dayPart]
      properties:
        id:
          type: string
          format: uuid
        user:
          type: object
          required: [id, name, colorCode]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            colorCode:
              type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dayPart:
          $ref: '#/components/schemas/DayPart'
        note:
          type: string
          nullable: true
          description: Always null in list responses. Populated in detail responses only for the PTO owner or a team lead.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Password:
      type: string
      format: password
      minLength: 8
      maxLength: 72
      description: |
        Plaintext password. Must be 8-72 characters (bcrypt's hard cap). Never
        echoed in any response body.
    SetupAccountRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 128
        password:
          $ref: '#/components/schemas/Password'
    CreateUserRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
          maxLength: 120
    CreateUserResponse:
      type: object
      required: [user, setupToken, expiresAt]
      properties:
        user:
          $ref: '#/components/schemas/User'
        setupToken:
          type: string
          description: |
            One-time plaintext setup token. Shown ONCE to the team lead and never
            recoverable from the server. The new member redeems it at
            /setup-account?token=...
        expiresAt:
          type: string
          format: date-time
    ResetPasswordResponse:
      type: object
      required: [setupToken, expiresAt]
      properties:
        setupToken:
          type: string
        expiresAt:
          type: string
          format: date-time
    CreatePTORequest:
      type: object
      description: Single-day PTO requires dayPart. Multi-day PTO is normalized to all_day. Weekend date selections and overlaps for the same user are rejected.
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dayPart:
          $ref: '#/components/schemas/DayPart'
        note:
          type: string
          maxLength: 500
      example:
        startDate: 2026-05-10
        endDate: 2026-05-10
        dayPart: morning
        note: Doctor appointment
    UpdatePTORequest:
      type: object
      description: Same validation rules as create. Update is allowed only for the owner or a team lead. Notes are visible only to the owner and team leads.
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dayPart:
          $ref: '#/components/schemas/DayPart'
        note:
          type: string
          maxLength: 500
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. Common values:
                  - `UNAUTHENTICATED` (401)
                  - `FORBIDDEN` (403)
                  - `VALIDATION_ERROR` (400)
                  - `NOT_FOUND` (404)
                  - `CONFLICT` (409)
                  - `RATE_LIMITED` (429) — see `Retry-After` header
                  - `NOT_READY` (503) — readiness probe only
                  - `INTERNAL_ERROR` (500)
              example: FORBIDDEN
            message:
              type: string
              example: You are not allowed to modify this PTO entry.